'Reporting' links go to tool-landing-page if no institution in request context.
When an institution is available, the 'Reporting' links in the top nav
and breadcrumbs still point to the tool-summary page for an
institution.  Sometimes, when a user has no STARS account, for
example, there is no institution available in the request context.
These changes handle those cases, by pointing the 'Reporting' links
to the tool-landing-page url.