The tool/menu.html 'My *' menu items only show when an instituion is available.
The 'My Summary', 'My Submission', 'My Resources', and 'My Reports(s)'
menu items are only shown when an instituion is available.  This is
almost always, but not quite; /tool/admin, for example, lacks an
institution.slug.

Since the 'My ...' menu items act on an institution, it doesn't hurt
to hide them when no institution is available.  In fact, errors will
be raised if '{% url ... instutition.slug %}' is called when no
institution is available.