Tweaked get_submit_url()'s in apps/submissions/models.py to handle inst slug.
get_submit_url() methods used to pass on the result of the get_submit_url()
method on a submission entity's counterpart in the credit hierarchy.  E.g.,
SubmissionSet.get_submit_url() used to return CreditSet.get_submit_url().

Now the URLs returned by the get_submit_url() methods in submissions.models
are calculated independent of the classes in credits.models.  These URLs
now include the slug of the related institution, which is available to
submission models, but not credit models.