Fixed submission bread crumbs.
Submission breadcrumbs used to link to get_submit_url() on credit
models.  Now they link to get_crumb_url() on the corresponding
submission models.

In most cases, get_crumb_url() returns self.get_submit_url().

get_submit_url() on the credit models now includes the institution and
submissionset info in the returned url, if a submissionset is passed
in.  The submission models' get_submit_url() wrap the corresponding
credit model's get_submit_url(), passing in the submission model's
submissionset.  This keeps the magic anchor syntax ("#ec_", etc.) in
one place (credits/models.py).

So, where it used to be:

   bread crumb url  <---  (credit)model.get_submit_url()

now it's:

   bread crumb url  <---  (submission)model.get_crumb_url()

Breadcrumb text used to be {{ model }} for model in credit models.
Now it's get_crumb_label() on the corresponding submission models.

On most submission models, get_crumb_label() returns the same as
{{ model }}, but it's neccesary because for SubmissionSet, {{ model }}
resolves to the name of the institution, plus the version of the
related CreditSet.  That seems redundant, so
SubmissionSet.get_crumb_label() returns just the CreditSet version.

So, where it used to be:

   bread crumb text  <---  {{ (credit)model }}

now it's:

   bread crumb text  <---  (submission)model.get_crumb_label()