Changes to cope with multiple inheritance.
Neccesary base class methods weren't being called via super() because
classes before them in the MRO didn't define the methods,
short-circuiting the flow.  Now the methods are there, and the derived
classes seem to work as is -- just don't mess around with the order of
the base classes in the class def. :-)