Error templates tweaks.
Added block 'user_message_block' to base_error.html -- gives 403.html
a slot to show content when {% if user_message %} is False.

Added block 'message_preamble' where '... use the menu ... to navigate
to the page you were looking for' message was.  Pushed that text into
the 'message_preamble' block defined in 404.html.  This keeps that
message from showing on 500's and 403's, too.