Rm'ed apps.institutions.tests.views.SubmissionInquiryViewTest.
Was written back when transition from flash messages to django
messages framework was made; just tested if messages show up when they
should.  Written more out of insecurity rather than necessity, and now
that changes to SubmissionInquiryView have broken it, it's time to
kill the test rather than fix it.