Subscription.purchase() turned into a class method.
purchase() now creates a Subscription, as well as purchasing it.
Putting the creation and purchase and possible payment in the same
place lets us roll back the creation of the subscription if the
payment fails (i.e., the credit card transaction is declined).

Before, one had to

    subscription = Subscription.create(...)
    Subscription.save()
    try:
        Subscription.purchase(...)
    except CreditCardProcessingError:
        subscription.delete()

Now it's just

    try:
        subscription = Subscription.purchase(...)
    except SubscriptionPurchaseError:
        ...

So that's better, right?  One need not worry about cleaning up if
purchase() fails; purchase() cleans up after itself.