Not using QuerySetSequence in AccountListView any more.
I copied QuerySetSequence from djangosnippets a while ago, and twice
it's been problematic, and now there's a bug in a -- let's say deeply
just to be dramatic -- nested lambda so screw it, it's outta there.

AccountListView used to be a ListView in which get_queryset() used
QuerySetSequence to chain a StarsAccount query and a PendingAccount
query.  Now it's a TemplateView in which get_context_data builds a
list of accounts from the same queries, and passes them to the view in
the template context.