StarsLiveServerTests can now run headlessly via phantomjs.
If phantomjs is installed (and in /usr/bin/local/bin (where
brew puts it)), it will be used.

If phantomjs is apparently not installed, Firefox will be
used.

If Firefox is apparently not installed, Chrome will be
used.  Note that in addition to the Google\ Chrome.app,
the chromedriver must be installed.  This is a separate
download.  http://code.google.com/p/chromedriver/downloads/list

Not yet tested on Linux.