Added support for automatic discounts.
Automatic discounts are ValueDiscounts with the new `automatic`
field set.

Effective date ranges of automatic discounts can't overlap,
so only one automatic ValueDiscount can be applicable at any
time.

Each automatic discount is applicable if today is in its
effective range, and its applicability filter is True.  An
applicability is a Python expression.  It can be as simple
as "True" so that the discount applies to all subscribers,
or more complicated, such as a discount that applies only
if a subscriber doesn't qualify for the early renewal
discount.

Applicability filters take two optional arguments, both
dictionaries.  The environment in which the filter expression
is evaluated is defined by adding `extra_globals` to globals(),
and `extra_locals` to locals().  This allows you to provide
any objects the filter expression requires.

The pricing view determines if there's an automatic discount
that should be applied, and passes this determination to the
pricing template

When an automatic discount is in effect, no other promo code
can be applied.

If an automatic discount should be applied, the pricing template
hides the promo code widgets, and applies the discount code.