Move Geordi in Requirements
Changed geordi from requirements_dev.txt to requirements.txt.

Still requires option "--process-dependency-links" to install, need
pip >=6.0.0 for this to be included inside requirements.txt. For now,
we can just run the whole pip install -r with that option and it
should work fine, but this is being deprecated with newer versions
of pip.