First File Based Caching Test
Added FileBased cache backend.

Added per-view caching in url patterns for institution reports as a proof
of concept test. If successful, this method will be expanded to other
problematic views next in a full audit of slow response times.

Duration of cache is set in-line with (86400 * number of days). This
can be changed at will or we could set this with an ENV variable we
can control in Ansible deployments; this first commit is just verifying
that this method is viable and does not have any unforseen consequences.