Template Fragment Caching
Switched tactics from per-view caching of the entire view to instead
creating a custom template tag with which we can choose an alternative
back-end and use these tags to cache specific segments of templates.

This required the installation of the django-adv-cache-tag package, and
the extension of the contained classes to implement a file-based cache
for a custom tag.

Tests were successful, it now caches /institutions/*/report/*/ pages
and serves the same version to all users BUT no longer caches the page
header bar containing user-specific information.

Next steps:
Additional arguments/settings to set cache max age, etc.
Versioning to cache separately for three user classes
The ability to manually invalidate a cache
Any relevant modifications to the ansible/gunicorn tasks