Cache Versioning Control
Modified template fragment caching by adding a version argument to
the custom tag. Counter-intuitively, if "versioning=True" in the class,
this will not work.

Introduced another custom tag that allows creation of a new varible within
the template. This was necessary to be able to perform the logic necessary
to differentiate the version when the tag is created...otherwise I had
to embed the entire page code inside the testrule tag for that version, again
inside the if tags when testing for staff, etc.

With this, you insert the same logic under the load tags in your desired
template, then add "version" within your cache tag. Done. Remove the testrule
tags if this is not an institution-specific page (i.e. testrule wasn't there
in the first place).

Added this modified fragment caching tag system to the relevant scripts
to test this versioning on the submission summary page and its sub-pages
for now for testing purposes.