Cache Invalidation Method
Needed to add the functionality to invalidate the cached version of reports
when a data correction is approved.

This required the cache key be more closely linked to the page being displayed,
otherwise it was impossible to find a specific report page and would have
required the purging of the entire cache directory.

The solution I devised was to manually set the cache key to be the URL of the
report page. This is redundant, but makes it easily retrievable with a manual
call to the cache API, and also allows me to reconstruct the key inside of
custom methods in other scripts without needing to invoke any methods of the
CacheTag() class outside of the context of the custom tag creation script.

To achieve this, I did the following:

1) Enable passing the request context through the template tag.

2) Pass the literal string 'request.get_full_path' as an argument to the tag
{% file_cache %} as the fragment name, as the template cannot be made to
resolve the variable's actual contained value inside the template itself.

3) Modify the get_base_cache_key() method of CacheTag to instead resolve that
literal string as a template variable using django.template.resolve_variable()
by passing the now accessible request context. This method is invoked when
CacheTag is instantiated and the cache is created, so this key modification
is inherited naturally AND can be invoked manually when making manual API calls
later on for invalidation.

4) Create a new invalidate_cache() method within the Data Correction model.
This makes an explicit call to the aforementioned get_base_cache_key() method
to construct the cache key from the present page's URL, looping through the
possible set of version identifiers, and invoking an API call to remove the
relevant cached file.

STILL NEEDED:

A method for invalidating the cache on the submission summary pages. These
are cached using additional arguments that make constructing the key much
more complex, in addition to needing to modify the current URL to reference
that one instead.

I may simplify the caching arguments on that summary page to facilitate this
if a solution is not readily apparent...just for the sake of time, it's not
worth fighting for a day and a half to invalidate it as is if I can just
alter how that key is created instead.

Also still working on expanding these cache tags to other templates, but was
waiting to finalize the tag formatting to do so.