Bug Fix - submissions.models.py
Method CategorySubmission.get_STARS_v2_0_score() had a bug. The final
statement before return expected a tuple, but the elif branch returned
values for total_achieved and total_available that always resulted in
the calculated score being an int.

Moved selection of the tuple value from innovation_score inside the if
branch and changed that final line to expect an int instead of a tuple.