Got Tests Passing
Got tetss for filecache template tag to pass using factories instead
of fixtures.

Modified institution factory to not use a float for the timestamp it
appends to the institution slug to avoid duplication, as the filecache
keys are delimited by dots. Simply changed this to an int to round off
the decimal.

Added an invalidation function to be invoked during SubmissionSet.save(),
and created an associated test which also passes.