Import BASIC_ACCESS in submissions.models
Where'd it go?  elpy-importmagic-fixup apparently dropped it.  That
really stinks.  So much for automatic refactoring.  Also, we should have
a test that fails when BASIC_ACCESS isn't imported -- I suppose that's
what 100% code coverage is about, and the value of tests that don't do
much more than execute the code.