Add migration (index on Institution.date_created)
Why a migration when the model hasn't changed?  Well, because the model
has changed.  The definition of Institution.date_created hasn't changed,
but Institution was defined twice in apps/institutions/models.py, and
the second definition -- without the 'db_index=True' notation -- was
recently removed.  Since the last one in wins, South never saw the
definition with 'db_index=True' until the second definition was
deleted.  Now that the second one is gone, South sees the first
definition, and recognizes the DB doesn't have a index on
Institution.date_created, so it creates a migration.