Update CreditUserSubmission.save
Now handles cases where submission_status tries to go from UNLOCKED to
IN_PROGRESS (this is disallowed).

Will send mail to stars-reviewers when CreditUserSubmissions change
submission_status from UNLOCKED to something else (e.g., COMPLETE) only
when the related SubmissionSet is under_review().  This allows us to
reset the submission_statuses from UNLOCKED to COMPLETE on all
CreditUserSubmissions related to a SubmissionSet when the SubmissionSet
is approved for publication.