Metric bug fix with populate button; performance improvements;
There are a few things going on here:

1. Fixed a bug where the imperial value was being copied for
institutions that preferred metric.
2. Removed `get_documentation_field_submission_value` in favor of
adding the context to the  `show_submission_form_for_field` and
`show_submission_form_for_field_inside_table` template tags. To avoid
duplicate calls for the value to be copied.
3. Improved performance when finding the related field using db queries
instead of a for loop on `get_submission_fields()`.
4. Changed the style of the button to include details about where the
value is being copied from and changed the style so it didn’t stand out
as much.
5. Consolidated duplicate template logic for the button into a single
file:`populate_button.html`.