Allow copying from submission data under review
Prior to this change, all input elements representing
DocumentationFieldSubmissions were disabled when the SubmissionSet is
under review.  Problem is you can't copy text out of a field that's
disabled, and reviewers have lots of text they want to copy out and
paste elsewhere.

Solution is to not set the disabled attribute of the text and textarea
elements, but set the readOnly (note camelCase) attribute instead.