Pass AccountListView.as_view() into never_cache
The account list view in manage/user is not updating after
a user is deleted. The page is loading very fast, so it
appears to be coming from a cache. By passing the view
into never_cache, I am trying to eliminate the confusion
by not letting the page be cached.